producer/requirements.txt
kafka-python==2.0.2
faker==8.14.0